debugging new offsets algorithm 1.1

+++ hw0samples/hw0a8cam2inst/hw0a8cam2inst.c
@@ -167,7 +167,7 @@ int main( int argc, char *argv[] )
               }
               printf(" : ");
 
-              offsets[bx] = offsets[bx]/left;
+              offsets[bx] = (int)offsets[bx]/(int)left;
 
               printf("%i\n", offsets[bx]);
